requirements.txt
streamlit
faiss-cpu
python-dotenv
google-genai
pypdf
pillow
sentence-transformers